Splitt of tokenizer into its own file
